Update README.md
I call this simple web app daily padi, it has all the futures to let you add the tasks you have to take care of, you can also set a reminder and also delete any task you wish to delete. when deleted all, it displays the message that says there is ni more task to show. it also has inbuilt ability to toggle the reminders to either on or off and a stylish + button that allows you to bring the adding dashboard to live and a - to close it back. The app is made with react.
the project is live at https://dailypadi.netlify.app/